capability: provide helpers for converting between xattrs and vfs_capability
Move the code in vfs_get_caps_from_disk() which converts an xattr to a
vfs_capability into a separate helper which will later be used for
getxattr/setxattr. Provide an equivalent helper for converting a
vfs_capability to an xattr, and some wrappers to hide the somewhat
confusing details of converting to/from userpsace.

Signed-off-by: Seth Forshee <sforshee@digitalocean.com>