capability: provide helpers for converting between xattrs and vfs_caps
vfs_get_caps_from_disk() contains code to convert a fscaps xattr to a
vfs_caps object. Separate this code out into a helper so it can be used
by other code.

Also add an analogous helper for converting a vfs_caps object to an
xattr. This helper is split into an internal interface, which does
not return an error if the rootid has no mapping in the destiation user
namespace, and an external interface which does return an error if the
rootid has no mapping. The internal interface will be used for
getxattr() to facilitate converting v3 capabilities to v2 capabilities
when the rootid corresponds to root in an ancestor user namesapce, and
the external interface will be used in all other cases.

Signed-off-by: Seth Forshee (DigitalOcean) <sforshee@kernel.org>